Looking like this is close to release 0.0.5 Updates to directions row, directions footer, realtimestoptask, getbuslocationtask
directionsrow.xml now has a hidden textview for light rail arrivals. This is default set to android:visibility=GONE
if the row is for a lightrail stop however, this will be made visible and a timertask will be assigned to update the row.

directionsfooter.xml essentially is updated to look prettier and text values changed. All on a single line now.

getbuslocationtask now checks is an infowindow is shown, if it is it hides and reshows it to show updates.